+++ src/commands/hardis/org/diagnose/legacyapi.ts
@@ -316,7 +316,8 @@ See article to solve issue before it's too late:
       let hostname;
       try {
         hostname = await dnsPromises.reverse(ip);
-      } catch (e) { // eslint-disable-line @typescript-eslint/no-unused-vars
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      } catch (e) {
         hostname = "unknown";
       }
       const ipResult = { CLIENT_IP: ip, CLIENT_HOSTNAME: hostname, SFDX_HARDIS_COUNT: ipInfo.count };