+++ src/commands/hardis/scratch/create.ts
@@ -331,7 +331,8 @@ export default class ScratchCreate extends SfdxCommand {
           await setConfig("user", {
             scratchOrgSfdxAuthUrl: this.scratchOrgSfdxAuthUrl,
           });
-        } catch (error) { // eslint-disable-line @typescript-eslint/no-unused-vars
+          // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        } catch (error) {
           uxLog(
             this,
             c.yellow(