+++ src/common/keyValueProviders/salesforce.ts
@@ -106,7 +106,8 @@ You mut create manually an Custom Object SfdxHardisKeyValueStore__c:
     // Initialize storage
     try {
       await setPoolStorage({}, options);
-    } catch (e) { // eslint-disable-line @typescript-eslint/no-unused-vars
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    } catch (e) {
       uxLog(this, c.yellow("You may have to create a Permission Set with all rights on SfdxHardisKeyValueStore__c and assign users to it"));
     }