+++ src/common/utils/index.ts
@@ -152,8 +152,10 @@ export async function checkAppDependency(appName) {
 }
 
 export async function promptInstanceUrl(orgTypes = ["login", "test"], alias = "default org", defaultOrgChoice: any = null) {
-  const customLoginUrlExample = orgTypes && orgTypes.length === 1 && orgTypes[0] === "login" ?
-    "https://myclient.lightning.force.com/" : "https://myclient--preprod.sandbox.lightning.force.com/";
+  const customLoginUrlExample =
+    orgTypes && orgTypes.length === 1 && orgTypes[0] === "login"
+      ? "https://myclient.lightning.force.com/"
+      : "https://myclient--preprod.sandbox.lightning.force.com/";
   const allChoices = [
     {
       title: "📝 Custom login URL (Sandbox, DevHub or Production Org)",
@@ -446,11 +448,11 @@ export async function interactiveGitAdd(options: any = { filter: [], groups: [] 
         this,
         c.grey(
           "The following list of files has not been proposed for selection\n" +
-          filesFiltered
-            .map((fileStatus: FileStatusResult) => {
-              return `  - (${getGitWorkingDirLabel(fileStatus.working_dir)}) ${getSfdxFileLabel(fileStatus.path)}`;
-            })
-            .join("\n"),
+            filesFiltered
+              .map((fileStatus: FileStatusResult) => {
+                return `  - (${getGitWorkingDirLabel(fileStatus.working_dir)}) ${getSfdxFileLabel(fileStatus.path)}`;
+              })
+              .join("\n"),
         ),
       );
     }
@@ -1152,7 +1154,8 @@ export async function generateSSLCertificate(branchName: string, folder: string,
       uxLog(commandThis, c.cyan(`Successfully deployed ${c.green(promptResponses.appName)} Connected App`));
       await fs.remove(tmpDirMd);
       await fs.remove(crtFile);
-    } catch (e) { // eslint-disable-line @typescript-eslint/no-unused-vars
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    } catch (e) {
       uxLog(
         commandThis,
         c.red("Error pushing ConnectedApp metadata. Maybe the app name is already taken ?\nYou may try again with another connected app name"),
@@ -1213,11 +1216,11 @@ export function getNested(nestedObj, pathArr) {
 }
 
 const ansiPattern = [
-  '[\\u001B\\u009B][[\\]()#;?]*(?:(?:(?:(?:;[-a-zA-Z\\d\\/#&.:=?%@~_]+)*|[a-zA-Z\\d]+(?:;[-a-zA-Z\\d\\/#&.:=?%@~_]*)*)?\\u0007)',
-  '(?:(?:\\d{1,4}(?:;\\d{0,4})*)?[\\dA-PR-TZcf-nq-uy=><~]))'
-].join('|');
-const ansiRegex = new RegExp(ansiPattern, 'g');
+  "[\\u001B\\u009B][[\\]()#;?]*(?:(?:(?:(?:;[-a-zA-Z\\d\\/#&.:=?%@~_]+)*|[a-zA-Z\\d]+(?:;[-a-zA-Z\\d\\/#&.:=?%@~_]*)*)?\\u0007)",
+  "(?:(?:\\d{1,4}(?:;\\d{0,4})*)?[\\dA-PR-TZcf-nq-uy=><~]))",
+].join("|");
+const ansiRegex = new RegExp(ansiPattern, "g");
 
 export function stripAnsi(str: string) {
-  return str.replace(ansiRegex, '');
+  return str.replace(ansiRegex, "");
 }